fixed issues with star a faction list of known systems not loading
I changed the Faction DB to save a list of star system Ids (which i had
to add). now instead of saving the whole star system, which would then
fail to load later, it saves just the id. at runtime this ID can be used
to lookup the full star system (if required, i suspect that most of the
time you just want to know if a faction is aware of the system, in that
case a simple guid vs guid check is all you need). There are two utility
functions to allow lookup of a star system one is in MiscHelper and is
the public API version, the other is in the game class and is internal
to the lib.